[stage] child is forking
This caused more mess to be created.

BUT, the manager node can actually listen to the queue while as
check if it's executing program has finsihed

When it does, send the AC_FP to master.

All of this generated managers to have states, which is variation of how
we listen to the queue.

Next step is to be able handle multiple execution requests considering the
time they were sent and also the delay