[fix] [critical] children not closing in shutdown
Children exit handler were modifying do_exit global variable on fork.c but
not in node.c, which is g_do_exit.

Children were never exiting, which is a serious process leak.

Solution is to create a function that is public that fork.c can access and tell
children to stop