[add] [feature] logging unit
- (add) (feature) logging unit that will help the develop debug certain
parts of the application without having to (un)comment lines of
printf.
The idea is to introduce groups of logging units that can be disabled
all at once by specifing it's name

Modules can have version by themselves. If we reach a state of stable
(thus trustworthy package), developers can remove the debugging assets.
If people want to rollback to debugging
tools, they can checkout to that verbose version of module.
Details can be reviews at issue #3

- (make) side note: be aware about the Makefile receipes.
A C module does not depend on the object, but the header file.
if implementation details change (i.e the c source file),
then the other object file shall remain the same
But if the interface changes (the header file), then the object file
(which is produced by compiling the c source file)
must be reviewed by the compiler.